fixed small warning

+++ YAHNC/YHNMasterViewController.m
@@ -129,7 +129,7 @@
     return headerView;
 }
 
--(float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
+-(CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
 {
     return 55.0;
 }
